requirements.txt
numpy==1.19.1
pandas==1.1.1
progressbar2==3.51.4
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
six==1.15.0
xlrd==1.1.0
